requirements.txt
requests~=2.25.1
pymongo~=3.11.4
Flask~=2.0.1
Werkzeug~=2.0.1
flask-login
flask-pymongo
flask-talisman
unidecode~=1.2.0
google-cloud-dialogflow
Pillow~=8.2.0
google~=3.0.0
pyIsEmail~=1.3.2
geocoder~=1.38.1